backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
numpy==1.26.4
opencv-python==4.9.0.80
deepface==0.0.91
tf-keras==2.21.0
ultralytics==8.3.0
python-dotenv==1.0.1
pytest==8.1.1
